ecomecre_v2/eCommerce_core/requirement.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
astroid==2.4.2
autopep8==1.5.5
bcrypt==3.1.7
blinker==1.4
Brlapi==0.7.0
certifi==2020.12.5
chardet==4.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.1
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.1.7
django-crispy-forms==1.11.0
django-heroku==0.3.1
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
fontawesome-free==5.15.2
future==0.18.2
gunicorn==20.0.4
httplib2==0.14.0
idna==2.10
isort==5.7.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
mailchimp-transactional==1.0.22
Mako==1.1.0
MarkupSafe==1.1.0
mccabe==0.6.1
monotonic==1.5
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==8.1.0
protobuf==3.6.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycairo==1.16.2
pycodestyle==2.6.0
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.6.2
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.8.1
python-debian===0.1.36ubuntu1
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.25.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
sqlparse==0.4.1
stripe==2.55.2
systemd-python==234
toml==0.10.2
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.3
usb-creator==0.3.7
virtualenv==20.4.1
wadllib==1.3.3
whitenoise==5.2.0
wrapt==1.12.1
xkit==0.0.0
